requirements.txt
protobuf
aiogram>=3.6.0,<4.0.0
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
google-cloud-firestore>=2.16.0
pydantic>=2.7.0

pytest>=8.3.3